requirements.txt
tensorflow == 2.16.1
numpy == 1.26.4
pandas == 2.2.2
fastapi == 0.111.0
pillow == 10.3.0 
pydantic == 2.7.1
uvicorn == 0.29.0